refactor: controller 클래스 패키지 이동 및 불필요한 코드 삭제
