+++ src/main/java/techcourse/myblog/web/Controller/IndexController.java
@@ -1,12 +1,10 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.Controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import techcourse.myblog.domain.repository.ArticleRepository;
-
-import javax.servlet.http.HttpSession;
 
 @Controller
 public class IndexController {
@@ -18,9 +16,8 @@ public class IndexController {
     }
 
     @GetMapping("/")
-    public String index(Model model, HttpSession httpSession) {
+    public String index(Model model) {
         model.addAttribute("articles", articleRepository.findAll());
-        model.addAttribute("user", httpSession.getAttribute("user"));
 
         return "index";
     }
