+++ src/main/java/techcourse/myblog/web/WebMvcConfig.java
@@ -3,7 +3,6 @@ package techcourse.myblog.web;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.util.PathMatcher;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -12,12 +11,11 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 public class WebMvcConfig implements WebMvcConfigurer {
     @Autowired
     @Qualifier(value = "authenticationInterceptor")
-    private HandlerInterceptor interceptor;
+    private HandlerInterceptor authenticationInterceptor;
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        PathMatcher patchMatcher;
-        registry.addInterceptor(interceptor)
+        registry.addInterceptor(authenticationInterceptor)
                 .addPathPatterns("/**");
     }
 }
